+++ providers/firebase/src/providers/data/slide.firebase.ts
@@ -1,59 +1,14 @@
-import firebase from 'firebase/app';
-import 'firebase/firestore';
-
 import {DeleteSlide, Slide, SlideData, UpdateSlide} from '@deckdeckgo/editor';
+import {deleteEntry, getEntry, updateEntry} from '../../utils/data/firestore.queries';
 
 export const getSlide = (deckId: string, slideId: string): Promise<Slide> => {
-  return new Promise<Slide>(async (resolve, reject) => {
-    const firestore: firebase.firestore.Firestore = firebase.firestore();
-
-    try {
-      const snapshot: firebase.firestore.DocumentSnapshot = await firestore.collection(`/decks/${deckId}/slides`).doc(slideId).get();
-
-      if (!snapshot.exists) {
-        reject('Slide not found');
-        return;
-      }
-
-      const slide: SlideData = snapshot.data() as SlideData;
-
-      resolve({
-        id: snapshot.id,
-        data: slide
-      });
-    } catch (err) {
-      reject(err);
-    }
-  });
+  return getEntry<SlideData>({id: slideId, collection: `/decks/${deckId}/slides`});
 };
 
-export const updateSlide: UpdateSlide = (deckId: string, slide: Slide): Promise<void> => {
-  return new Promise<void>(async (resolve, reject) => {
-    const firestore: firebase.firestore.Firestore = firebase.firestore();
-
-    const now: firebase.firestore.Timestamp = firebase.firestore.Timestamp.now();
-    slide.data.updated_at = now as unknown as Date;
-
-    try {
-      await firestore.collection(`/decks/${deckId}/slides`).doc(slide.id).set(slide.data, {merge: true});
-
-      resolve();
-    } catch (err) {
-      reject(err);
-    }
-  });
+export const updateSlide: UpdateSlide = (deckId: string, slide: Slide): Promise<Slide> => {
+  return updateEntry<Slide>({entry: slide, collection: `/decks/${deckId}/slides`});
 };
 
 export const deleteSlide: DeleteSlide = (deckId: string, slideId: string): Promise<void> => {
-  return new Promise<void>(async (resolve, reject) => {
-    try {
-      const firestore: firebase.firestore.Firestore = firebase.firestore();
-
-      await firestore.collection(`/decks/${deckId}/slides`).doc(slideId).delete();
-
-      resolve();
-    } catch (err) {
-      reject(err);
-    }
-  });
+  return deleteEntry({id: slideId, collection: `/decks/${deckId}/slides`});
 };