+++ studio/src/app/pages/editor/app-editor.tsx
@@ -1,10 +1,16 @@
 import {Component, ComponentInterface, Fragment, h, Listen, State} from '@stencil/core';
 
 import {isIOS} from '@deckdeckgo/utils';
+import {SyncEvent} from '@deckdeckgo/editor';
 
 import {Editor} from '../../types/editor/editor';
 
 import {getEdit} from '../../utils/editor/editor.utils';
+
+import {initSyncState, sync} from '../../providers/sync/sync.provider';
+
+import {worker} from '../../workers/sync.worker.ts?worker';
+import {startSyncTimer, stopSyncTimer} from '../../workers/sync.worker';
 
 @Component({
   tag: 'app-editor',
@@ -28,6 +34,28 @@ export class AppDeckEditor implements ComponentInterface {
     } else if ($event?.detail?.from === '/') {
       await this.destroy();
     }
+  }
+
+  async componentDidLoad() {
+    await this.syncData();
+  }
+
+  async disconnectedCallback() {
+    await stopSyncTimer();
+  }
+
+  private async syncData() {
+    await startSyncTimer();
+
+    worker.onmessage = async ({data}: MessageEvent<SyncEvent>) => {
+      if (!data || data.msg !== 'deckdeckgo_sync') {
+        return;
+      }
+
+      await sync(data.data);
+    };
+
+    await initSyncState();
   }
 
   private async init() {
@@ -61,7 +89,7 @@ export class AppDeckEditor implements ComponentInterface {
   render() {
     return (
       <Fragment>
-        <app-navigation class={this.hideNavigation ? 'hidden' : undefined}></app-navigation>
+        <app-navigation actions="all" class={this.hideNavigation ? 'hidden' : undefined}></app-navigation>
 
         {this.renderEditor()}
       </Fragment>