+++ providers/firebase/CHANGELOG.md
@@ -1,3 +1,15 @@
+# 1.1.1 (2021-11-07)
+
+### Build
+
+- published editor types as dev dependency
+
+# 1.1.0 (2021-11-07)
+
+### Features
+
+- documents and paragraphs supports
+
 # 1.0.1 (2021-10-21)
 
 ### Fix